Moved agave.modules.nic to agave.nic, and added main to module
