+++ agave/arp/resolve.py
@@ -6,7 +6,7 @@ from typing import Union
 from .utils import _create_socket, _parse, SOCKET_PROTO, SOCKET_MAX_READ
 from agave.frames.ethernet import MACAddress
 from agave.frames import ethernet, arp
-from agave.modules.nic.interfaces import NetworkInterface
+from agave.nic.interfaces import NetworkInterface
 from ipaddress import IPv4Address
