+++ agave/nic/interfaces.py
@@ -176,3 +176,42 @@ class NetworkInterface:
         raise NetworkInterfaceNotFound(
             "No interface found connected to a network including %s" % (host)
         )
+
+if __name__ == "__main__":
+    """
+    Prints information about network interfaces.
+
+    Usage:
+        python3 -m agave.nic.interfaces [interface]
+    
+    Examples:
+        python3 -m agave.nic.interfaces
+        python3 -m agave.nic.interfaces eth0
+
+    """
+    import sys
+
+
+    if len(sys.argv) > 1:
+        try:
+            nic = NetworkInterface.get_by_name(sys.argv[1])
+        except NetworkInterfaceNotFound as e:
+            print(e)
+        else:
+            print("{:20}{}\n{:20}{}\n{:20}{}\n{:20}{}\n{:20}{}".format(
+                "NAME", nic.name,
+                "MAC", nic.mac,
+                "IP", nic.ip,
+                "NETWORK", nic.network,
+                "BROADCAST", nic.broadcast
+            ))
+    else:
+        print("{:20}\t{:20}\t{:20}\t{:20}\t".format("NAME", "MAC", "IP", "NETWORK"))
+        print("".join(["-"] * 88))
+        nics = NetworkInterface.list()
+        for i in range(0, len(nics)):
+            nic = nics[i]
+            print("{:20}\t{:20}\t{:20}\t{:20}\t".format(
+                nic.name, str(nic.mac), str(nic.ip), str(nic.network)
+            ))
+
